requirements.txt
opencv-python
numpy<2.0
deepface
faiss-cpu
tf-keras
fastapi
uvicorn
python-multipart
sarvamai
python-dotenv
pyaudio
vosk
pydub